[test] apply the new test style to common-mem-allocator
