+++ common/mem-allocator/Cargo.toml
@@ -5,6 +5,10 @@ edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[lib]
+doctest = false
+test = false
+
 [dependencies]
 common-base = {path = "../base" }
 tikv-jemalloc-sys = "0.4.2"
